web-audio/gain-node.html

<!doctype html>
<html class="a">
 <head>
  <title>GainNode interface</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="js/lodash.js"></script>
  <script src="js/vendor-prefixes.js"></script>
  <script src="js/helpers.js"></script>
 </head>
 <body class="a">
  <div id="log"></div>
  <script>

var gainNode;
setup(function() {
  context = new AudioContext();
  gainNode = context.createGain();
});

test(function() {
  assert_equals(gainNode.numberOfInputs, 1, "gainNode.numberOfInputs");
}, "gainNode.numberOfInputs should be 1 by default");

test(function() {
  assert_equals(gainNode.numberOfOutputs, 1, "gainNode.numberOfOutputs");
}, "gainNode.numberOfOutputs should be 1 by default");

test(function() {
  assert_has_property(gainNode, "gain", "gainNode should have gain");
}, "gainNode should have gain");

test_implements_audio_param_interface(gainNode.gain);

  </script>
 </body>
</html>
